+++ packages/krtour-map-admin/src/krtour/map_admin/routers/__init__.py
@@ -33,6 +33,7 @@ from krtour.map_admin.routers.features import (
 from krtour.map_admin.routers.mois_detail import router as mois_detail_router
 from krtour.map_admin.routers.offline_uploads import router as offline_uploads_router
 from krtour.map_admin.routers.ops import router as ops_router
+from krtour.map_admin.routers.ops_live import router as ops_live_router
 from krtour.map_admin.routers.ops_logs import router as ops_logs_router
 from krtour.map_admin.routers.poi_cache_targets import (
     router as poi_cache_targets_router,
@@ -55,6 +56,7 @@ __all__ = [
     "mois_detail_router",
     "offline_uploads_router",
     "ops_router",
+    "ops_live_router",
     "ops_logs_router",
     "dagster_router",
     "public_status_router",